third_party/WebKit/LayoutTests/animations/element-animate-no-style-recalc.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<script>
test(() => {
  internals.updateStyleAndReturnAffectedElementCount();
  container.style.color = 'blue';
  target.animate({color: 'blue'}, {fill: 'forwards'});
  assert_greater_than(internals.updateStyleAndReturnAffectedElementCount(), 1);
}, 'Calling element.animate() should not force a style recalc.');
</script>
